+++ src/generator/resources/regmap.py
@@ -2,44 +2,46 @@ from collections.abc import Generator
 from enum import IntEnum
 import typing
 
+
 # pyright: reportAny = false
 
-class Peripheral(object):
-    '''address and collection of registers'''
 
-    def __init__(self, name: str = '', addr: int = 0) :
-        self.__addr__ : int = addr
-        self.__name__ : str = name
+class Peripheral(object):
+    """address and collection of registers"""
+
+    def __init__(self, name: str = '', addr: int = 0):
+        self.__addr__: int = addr
+        self.__name__: str = name
 
     def rifs(self) -> Generator['Peripheral']:
-        '''Iterator over all rifs'''
+        """Iterator over all rifs"""
         for attr_name in vars(self):
             obj = getattr(self, attr_name)
             if isinstance(obj, Peripheral):
                 yield obj
 
     def regs(self) -> Generator['Register']:
-        '''Iterator over all registers'''
+        """Iterator over all registers"""
         for attr_name in vars(self):
             obj = getattr(self, attr_name)
             if isinstance(obj, Register):
                 yield obj
 
     def address(self) -> int:
-        '''Address of the peripheral'''
+        """Address of the peripheral"""
         return self.__addr__
 
     def name(self) -> str:
-        '''Instance name of the peripheral'''
+        """Instance name of the peripheral"""
         return self.__name__
 
     def by_name(self, path: str) -> 'None|Peripheral|Register|Field':
-        '''Retrieve an element or sub-element of the peripheral by name
-        Accept path such as 'rif.reg.field' '''
+        """Retrieve an element or sub-element of the peripheral by name
+        Accept path such as 'rif.reg.field' """
         obj = self
         for n in path.split('.'):
             o = getattr(obj, n)
-            if isinstance(o, (Peripheral,Register,Field)):
+            if isinstance(o, (Peripheral, Register, Field)):
                 obj = o
             else:
                 return None
@@ -47,7 +49,7 @@ class Peripheral(object):
 
 
 class Register(object):
-    '''Register definition: address and collection of fields'''
+    """Register definition: address and collection of fields"""
 
     @typing.final
     class regInfo:
@@ -57,10 +59,10 @@ class Register(object):
             self.name = name
             self.address = address
             self.readonly = readonly
-            self.flags : list[str] = []
+            self.flags: list[str] = []
 
-    def __init__(self, parent: None | Peripheral, name: str, addr: int, readonly: bool, init: None|int = None) :
-        self.__reg_info__ : Register.regInfo = Register.regInfo(parent, name, addr, readonly)
+    def __init__(self, parent: None | Peripheral, name: str, addr: int, readonly: bool, init: None | int = None):
+        self.__reg_info__: Register.regInfo = Register.regInfo(parent, name, addr, readonly)
         if init is not None:
             self.set_value(init)
 
@@ -72,7 +74,7 @@ class Register(object):
         return s
 
     def fields(self):
-        '''Iterator over all fields'''
+        """Iterator over all fields"""
         for attr_name in vars(self):
             obj = getattr(self, attr_name)
             if isinstance(obj, Field):
@@ -83,61 +85,62 @@ class Register(object):
                         yield x
 
     def parent(self) -> None | Peripheral:
-        '''Return peripheral which owns the register'''
+        """Return peripheral which owns the register"""
         return self.__reg_info__.parent
 
     def name(self) -> str:
-        '''Instance name of the register'''
+        """Instance name of the register"""
         return self.__reg_info__.name
 
     def readonly(self) -> bool:
-        '''Return true when register is read-only'''
+        """Return true when register is read-only"""
         return self.__reg_info__.readonly
 
     def flags(self) -> list[str]:
-        '''Register flags: interrupt, external, ...'''
+        """Register flags: interrupt, external, ..."""
         return self.__reg_info__.flags
 
     def by_name(self, name: str) -> 'None|Field':
-        '''Retrieve a field by name'''
+        """Retrieve a field by name"""
         obj = getattr(self, name)
         return obj if isinstance(obj, Field) else None
 
     def address(self) -> int:
-        '''Absolute address of the register'''
+        """Absolute address of the register"""
         if self.__reg_info__.parent is not None:
             return self.__reg_info__.parent.__addr__ + self.__reg_info__.address
         return self.__reg_info__.address
 
     def set_value(self, value_all_fields: int):
-        '''Set value for the register'''
+        """Set value for the register"""
         for field in self.fields():
             reg_val = value_all_fields >> field.pos
-            mask = (1<<field.width) - 1
+            mask = (1 << field.width) - 1
             reg_val &= mask
             field.value = reg_val
 
     def get_value(self) -> int:
-        '''Get value for the register'''
+        """Get value for the register"""
         reg_val = 0
         for field in self.fields():
             mask = (1 << field.width) - 1
             reg_val |= (field.value & mask) << field.pos
         return reg_val
 
-class Field(object):
-    '''field value, width and position inside a register'''
 
-    value  : int = 0
-    width  : int = 0
-    pos    : int = 0
-    name   : str = ""
-    kind   : str = ""
-    signed : bool = False
+class Field(object):
+    """field value, width and position inside a register"""
+
+    value: int = 0
+    width: int = 0
+    pos: int = 0
+    name: str = ""
+    kind: str = ""
+    signed: bool = False
     nb_frac: int = 0
 
-    def __init__(self, parent: Register, name: str = '', init: None|int = None) :
-        self.__parent__ : Register = parent
+    def __init__(self, parent: Register, name: str = '', init: None | int = None):
+        self.__parent__: Register = parent
         self.name = name
         if init is not None:
             self.value = init
@@ -151,52 +154,52 @@ class Field(object):
         else:
             if self.nb_frac != 0:
                 s = f'{self.get_float()} ({v})'
-            else :
+            else:
                 s = f'{v}'
         return s
 
     def parent(self) -> Register:
-        '''Return register owning the field'''
+        """Return register owning the field"""
         return self.__parent__
 
     def mask(self) -> int:
-        '''Return mask corresponding to the field in the register'''
-        return ((1<<self.width) - 1) << self.pos
+        """Return mask corresponding to the field in the register"""
+        return ((1 << self.width) - 1) << self.pos
 
     def set_value(self, value: int | float):
-        '''Set the field value, either as internal integer representation or the corresponding fixed-point value'''
+        """Set the field value, either as internal integer representation or the corresponding fixed-point value"""
         if isinstance(value, int):
             self.value = value
-        else :
+        else:
             conv = getattr(self.enum_kind(), 'from_float', None)
             if conv is not None:
                 self.value = conv(value)
             elif self.nb_frac < 0:
                 self.value = int(value / (1 << (-self.nb_frac)))
-            else :
+            else:
                 self.value = int(value * (1 << self.nb_frac))
 
     def get_value(self) -> int:
-        '''Return the field value (internal integer representation)'''
+        """Return the field value (internal integer representation)"""
         v = self.value
-        if self.signed and v > 1<<(self.width-1):
-            v -= (1<<self.width)
+        if self.signed and v > 1 << (self.width - 1):
+            v -= (1 << self.width)
         return v
 
     def get_float(self) -> float:
-        '''Return the field value (fixed-point representation)'''
+        """Return the field value (fixed-point representation)"""
         v = float(self.get_value())
         if self.nb_frac < 0:
             v = v * (1 << (-self.nb_frac))
-        else :
+        else:
             v = v / (1 << self.nb_frac)
         return v
 
     def address(self) -> int:
-        '''Return address of the register owning the field'''
-        if self.__parent__ :
+        """Return address of the register owning the field"""
+        if self.__parent__:
             return self.__parent__.address()
         return 0
 
-    def enum_kind(self) -> None| type[IntEnum]:
+    def enum_kind(self) -> None | type[IntEnum]:
         return None